shoping cart is developping

+++ backend/kukorica/src/main/java/hu/back/kukorica/dao/UserRepository.java
@@ -7,7 +7,8 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface UserRepository extends JpaRepository<User,Long> {
+public interface UserRepository extends JpaRepository<User,Integer> {
 
     User findByEmail(String email);
+
 }
